Prevent return of undefined from parser.
If the user enters a value into the Auto-Complete widget then clears it
out again it can cause the parser to return an undefined result.  This,
in turn, marks the parser as not valid and applies the ng-invalid class
to the widget.  The solution is to initialize the returned results
variable.

See $$parseAndValidate (angular.js 716)